Refactored MenuItemPage componenent and created a new folder to store all menu components together. Added functionality to sides menu items.
